backend/requirements.txt
fastapi==0.100.0
pydantic==1.10.12
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.20
psycopg2-binary==2.9.11
alembic==1.12.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
requests==2.31.0
meilisearch==0.37.1
beautifulsoup4==4.12.2
lxml==4.9.3
jdatetime==3.0.0
reportlab==4.0.0
openpyxl==3.1.2
pandas==2.2.2
cryptography==41.0.3
email-validator==2.1.0
python-multipart==0.0.6
pyotp==2.9.0
qrcode==8.2
